Property Changes Property Change Notifications: The OnPropertyChanged method is used to notify the view when any bound property changes (like Number1, Number2, or Result). This reverts the previous way of updating the UI when the numbers were changed
Commands Redo
OperationCommand is triggered when an operation button (Add, Subtract, etc.) is pressed. It sets the current operation and switches the input to Number2.

EqualsCommand executes the stored operation and updates the result when the Equals button is pressed.

ClearCommand resets the calculator inputs and clears the operation.

NumberButtonCommand appends the clicked number to either Number1 or Number2 depending on which one is active.

The Reset method is used to clear the inputs and reset the state of the calculator after an operation is performed.